requirements.txt
torch>=2.0.0
transformers>=4.40.0
datasets>=2.14.0
accelerate>=0.25.0
peft>=0.7.0
trl>=0.7.0
bitsandbytes>=0.41.0
qwen-vl-utils>=0.0.8
vllm>=0.4.0
pytrec_eval-terrier
tqdm
ftfy
opencv-python
Pillow
numpy
